# internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: managed-agent-go
  version: 0.1.0
  description: >-
    Clean-room implementation of the supported Claude Managed Agents API core,
    conditional Files and Memory APIs, and File- and Memory-backed Session
    Resources. This project is Apache-2.0 licensed and is not an Anthropic
    product.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Current managed-agent-go server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
...
paths:
  /v1/agents:
    post:
      operationId: createAgent
      tags: [Agents]
      summary: Create an Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listAgents
      tags: [Agents]
      summary: List Agents
      parameters:
        - $ref: '#/components/parameters/AgentListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/IncludeArchived'
        - $ref: '#/components/parameters/CreatedAtGte'
        - $ref: '#/components/parameters/CreatedAtLte'
      responses:
        '200':
          $ref: '#/components/responses/AgentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    get:
      operationId: getAgent
      tags: [Agents]
      summary: Get an Agent
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: updateAgent
      tags: [Agents]
      summary: Update an Agent and create a new version when fields change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/agents/{agent_id}/versions:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    get:
      operationId: listAgentVersions
      tags: [Agents]
      summary: List Agent versions
      parameters:
        - $ref: '#/components/parameters/AgentListLimit'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/AgentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/agents/{agent_id}/archive:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    post:
      operationId: archiveAgent
      tags: [Agents]
      summary: Archive an Agent
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/environments:
    post:
      operationId: createEnvironment
      tags: [Environments]
      summary: Create an Environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listEnvironments
      tags: [Environments]
      summary: List Environments
      parameters:
        - $ref: '#/components/parameters/EnvironmentListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/IncludeArchived'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /v1/environments/{environment_id}:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    get:
      operationId: getEnvironment
      tags: [Environments]
      summary: Get an Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: updateEnvironment
      tags: [Environments]
      summary: Update an Environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteEnvironment
      tags: [Environments]
      summary: Delete an unreferenced Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/environments/{environment_id}/archive:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    post:
      operationId: archiveEnvironment
      tags: [Environments]
      summary: Archive an Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/files:
    post:
      operationId: uploadFile
      tags: [Files]
      summary: Upload a File
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/FileResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listFiles
      tags: [Files]
      summary: List Files
      parameters:
        - $ref: '#/components/parameters/FileListLimit'
        - $ref: '#/components/parameters/AfterID'
        - $ref: '#/components/parameters/BeforeID'
        - $ref: '#/components/parameters/ScopeID'
      responses:
        '200':
          $ref: '#/components/responses/FileListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/files/{file_id}:
    parameters:
      - $ref: '#/components/parameters/FileID'
    get:
      operationId: getFileMetadata
      tags: [Files]
      summary: Get File metadata
      responses:
        '200':
          $ref: '#/components/responses/FileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Delete a File
      responses:
        '200':
          $ref: '#/components/responses/FileDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/files/{file_id}/content:
    parameters:
      - $ref: '#/components/parameters/FileID'
    get:
      operationId: downloadFile
      tags: [Files]
      summary: Download a File marked as downloadable
      responses:
        '200':
          $ref: '#/components/responses/FileContentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills:
    post:
      operationId: createSkill
      tags: [Skills]
      summary: Create a custom Skill and its first immutable Version
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SkillUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/SkillResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listSkills
      tags: [Skills]
      summary: List Skills
      parameters:
        - $ref: '#/components/parameters/SkillListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/SkillSource'
      responses:
        '200':
          $ref: '#/components/responses/SkillListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills/{skill_id}:
    parameters:
      - $ref: '#/components/parameters/SkillID'
    get:
      operationId: getSkill
      tags: [Skills]
      summary: Get a Skill
      responses:
        '200':
          $ref: '#/components/responses/SkillResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSkill
      tags: [Skills]
      summary: Delete a Skill after all of its Versions have been deleted
      responses:
        '200':
          $ref: '#/components/responses/SkillDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills/{skill_id}/versions:
    parameters:
      - $ref: '#/components/parameters/SkillID'
    post:
      operationId: createSkillVersion
      tags: [Skills]
      summary: Create an immutable Skill Version
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SkillVersionUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listSkillVersions
      tags: [Skills]
      summary: List immutable Skill Versions
      parameters:
        - $ref: '#/components/parameters/SkillVersionListLimit'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills/{skill_id}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/SkillID'
      - $ref: '#/components/parameters/SkillVersion'
    get:
      operationId: getSkillVersion
      tags: [Skills]
      summary: Get a Skill Version
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSkillVersion
      tags: [Skills]
      summary: Delete a Skill Version and its archived contents
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills/{skill_id}/versions/{version}/content:
    parameters:
      - $ref: '#/components/parameters/SkillID'
      - $ref: '#/components/parameters/SkillVersion'
    get:
      operationId: downloadSkillVersion
      tags: [Skills]
      summary: Download a Skill Version as a zip archive
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionContentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
    post:
      operationId: createMemoryStore
      tags: [Memory]
      summary: Create a Memory Store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryStoreCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listMemoryStores
      tags: [Memory]
      summary: List Memory Stores
      parameters:
        - $ref: '#/components/parameters/MemoryStoreListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/IncludeArchived'
        - $ref: '#/components/parameters/CreatedAtGte'
        - $ref: '#/components/parameters/CreatedAtLte'
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
    get:
      operationId: getMemoryStore
      tags: [Memory]
      summary: Get a Memory Store
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    post:
      operationId: updateMemoryStore
      tags: [Memory]
      summary: Update a Memory Store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryStoreUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteMemoryStore
      tags: [Memory]
      summary: Delete a Memory Store
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/archive:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
    post:
      operationId: archiveMemoryStore
      tags: [Memory]
      summary: Archive a Memory Store
      responses:
        '200':
          $ref: '#/components/responses/MemoryStoreResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/memories:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
    post:
      operationId: createMemory
      tags: [Memory]
      summary: Create a Memory
      parameters:
        - $ref: '#/components/parameters/MemoryView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/MemoryResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listMemories
      tags: [Memory]
      summary: List Memories and optional depth-one path prefixes
      parameters:
        - $ref: '#/components/parameters/MemoryListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/MemoryView'
        - $ref: '#/components/parameters/MemoryPathPrefix'
        - $ref: '#/components/parameters/MemoryDepth'
      responses:
        '200':
          $ref: '#/components/responses/MemoryListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/memories/{memory_id}:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
      - $ref: '#/components/parameters/MemoryID'
    get:
      operationId: getMemory
      tags: [Memory]
      summary: Get a Memory
      parameters:
        - $ref: '#/components/parameters/MemoryView'
      responses:
        '200':
          $ref: '#/components/responses/MemoryResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    post:
      operationId: updateMemory
      tags: [Memory]
      summary: Update a Memory with an optional SHA-256 precondition
      parameters:
        - $ref: '#/components/parameters/MemoryView'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/MemoryResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteMemory
      tags: [Memory]
      summary: Delete a Memory with an optional SHA-256 precondition
      parameters:
        - $ref: '#/components/parameters/ExpectedContentSHA256'
      responses:
        '200':
          $ref: '#/components/responses/MemoryDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/memory_versions:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
    get:
      operationId: listMemoryVersions
      tags: [Memory]
      summary: List immutable Memory Versions
      parameters:
        - $ref: '#/components/parameters/MemoryVersionListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/MemoryView'
        - $ref: '#/components/parameters/MemoryAPIKeyIDFilter'
        - $ref: '#/components/parameters/MemorySessionIDFilter'
        - $ref: '#/components/parameters/MemoryIDFilter'
        - $ref: '#/components/parameters/MemoryOperationFilter'
        - $ref: '#/components/parameters/CreatedAtGte'
        - $ref: '#/components/parameters/CreatedAtLte'
      responses:
        '200':
          $ref: '#/components/responses/MemoryVersionListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/memory_versions/{version_id}:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
      - $ref: '#/components/parameters/MemoryVersionID'
    get:
      operationId: getMemoryVersion
      tags: [Memory]
      summary: Get an immutable Memory Version
      parameters:
        - $ref: '#/components/parameters/MemoryView'
      responses:
        '200':
          $ref: '#/components/responses/MemoryVersionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/memory_stores/{store_id}/memory_versions/{version_id}/redact:
    parameters:
      - $ref: '#/components/parameters/MemoryBeta'
      - $ref: '#/components/parameters/MemoryStoreID'
      - $ref: '#/components/parameters/MemoryVersionID'
    post:
      operationId: redactMemoryVersion
      tags: [Memory]
      summary: Permanently redact the content of an historical Memory Version
      responses:
        '200':
          $ref: '#/components/responses/MemoryVersionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    AgentID:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    AgentListLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    EnvironmentListLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
...
    Page:
      name: page
      in: query
      description: Opaque cursor returned by a previous list response.
      schema:
        type: string
        minLength: 1
    IncludeArchived:
      name: include_archived
      in: query
      schema:
        type: boolean
        default: false
...
    CreatedAtGte:
      name: created_at[gte]
      in: query
      schema:
        type: string
        format: date-time
...
    CreatedAtLte:
      name: created_at[lte]
      in: query
      schema:
        type: string
        format: date-time
  responses:
    AgentResponse:
      description: Agent resource
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Agent'
    AgentListResponse:
      description: Agent page
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AgentList'
...
    InvalidRequest:
      description: Invalid request
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnsupportedRequest:
      description: Well-formed capability that this implementation cannot enforce
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    AgentCreateRequest:
      type: object
      additionalProperties: false
      required: [name, model]
      properties:
        name:
          type: string
          minLength: 1
        model:
          $ref: '#/components/schemas/ModelInput'
        system:
          type: [string, 'null']
        description:
          type: [string, 'null']
        tools:
          type: array
          maxItems: 128
          items:
            $ref: '#/components/schemas/AgentTool'
        mcp_servers:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/MCPServer'
        skills:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/SkillReferenceInput'
        multiagent:
          $ref: '#/components/schemas/Multiagent'
        metadata:
          $ref: '#/components/schemas/Metadata'
    AgentUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        version:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
        model:
          $ref: '#/components/schemas/ModelInput'
        system:
          type: [string, 'null']
        description:
          type: [string, 'null']
        tools:
          oneOf:
            - type: array
              maxItems: 128
              items:
                $ref: '#/components/schemas/AgentTool'
            - type: 'null'
        mcp_servers:
          oneOf:
            - type: array
              maxItems: 20
              items:
                $ref: '#/components/schemas/MCPServer'
            - type: 'null'
        skills:
          oneOf:
            - type: array
              maxItems: 500
              items:
                $ref: '#/components/schemas/SkillReferenceInput'
            - type: 'null'
        multiagent:
          $ref: '#/components/schemas/Multiagent'
        metadata:
          $ref: '#/components/schemas/MetadataPatch'
...